cleanup api using namesapces and add hyper parameter container
