+++ src/dmme/ddpm/__init__.py
@@ -1,7 +1,7 @@
-from .training import TrainState
-from .training import LinearSchedule
-from .training import forward_process
-from .training import reverse_process
-from .training import train_step
+from . import train
+from . import sample
+from . import schedule
 
-from .models import UNet
+from . import models
+
+from . import data
