+++ src/dmme/ddpm/data.py
@@ -1,0 +1,2 @@
+def norm(x):
+    return (x - 0.5) * 2
